Update issue 5924 Oops, if onBeforeRequest's block-results are sent before beforeload fires, I was always deleting the element, even if it wasn't blocked.
svn path=/branches/issue-5924/; revision=3002

+++ src/adblock_start_new_style.js
@@ -31,7 +31,8 @@ var mightRemove = {
         delete mightRemove.sentEarly[key];
         log("Deleted early block result key", key);
       }
-      destroyElement(event.target, elType);
+      if (early.blocked)
+        destroyElement(event.target, elType);
     }
     else {
       mightRemove.add(key, event.target);
